database modifiactions + flyway installation

+++ src/main/java/com/reit/reit/respositry/ActivityRepository.java
@@ -3,5 +3,5 @@ package com.reit.reit.respositry;
 import com.reit.reit.model.Activity;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface ActivityRepository extends PagingAndSortingRepository<Activity, Long> {
+public interface ActivityRepository extends PagingAndSortingRepository<Activity, Integer> {
 }
