+++ src/main/java/com/reit/reit/respositry/BuildingRepository.java
@@ -3,5 +3,5 @@ package com.reit.reit.respositry;
 import com.reit.reit.model.Building;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface BuildingRepository extends PagingAndSortingRepository<Building, Long> {
+public interface BuildingRepository extends PagingAndSortingRepository<Building, Integer> {
 }
