+++ src/main/java/com/reit/reit/respositry/UserRepository.java
@@ -3,5 +3,5 @@ package com.reit.reit.respositry;
 import com.reit.reit.model.User;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
-public interface UserRepository extends PagingAndSortingRepository<User, Long> {
+public interface UserRepository extends PagingAndSortingRepository<User, Integer> {
 }
